requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
csv342==1.0.1
google-api-core==2.11.0
google-auth==2.16.1
google-cloud==0.34.0
google-cloud-bigquery==3.6.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
jq==1.4.0
mypy==1.0.1
mypy-extensions==1.0.0
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
requests==2.28.2
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
